Move files related to tokenizing to separate directory
