+++ include/flang/pp/print_token.hpp
@@ -3,7 +3,7 @@
 #include <iostream>
 #include <vector>
 
-#include "flang/token.hpp"
+#include "flang/tokenize/token.hpp"
 
 namespace flang {
 std::ostream& operator<<(std::ostream& os, Token const& token);
