+++ src/flang/tokenize/tokenizer.cpp
@@ -1,4 +1,4 @@
-#include "flang/tokenizer.hpp"
+#include "flang/tokenize/tokenizer.hpp"
 
 #include "flang/flang_exception.hpp"
